Changement dre prix en double

+++ HomeCinema-ejb/src/java/ejbs/admin/ManageProduct.java
@@ -35,7 +35,7 @@ public class ManageProduct implements ManageProductRemote {
     public EntityManager em;
 
     @Override
-    public Long createProductWithFilm(FilmDto fdto, VideoDto trailer, VideoDto vid, Integer price) {
+    public Long createProductWithFilm(FilmDto fdto, VideoDto trailer, VideoDto vid, Double price) {
         Film f = ManageEntitieFilm.createFilmWithVideo(fdto, trailer, vid, em);
         Product p = new Product(f, price);
         f.setMain_product(p);
